chore: cleanup
This removes a bunch of comments and moves functions to the appropriate
services and replaces existing functions with dummy data with real
queries to firestore.
