+++ src/fakeData/facilities.ts
@@ -1,46 +1,6 @@
-import { HealthcareProfessional, Locale, LocaleName, Degree, 
-    Specialty, SpecialtyName, SpokenLanguage, Insurance, Facility } from '../typeDefs/dbSchema'
+import { Facility } from '../typeDefs/dbSchema'
 
 export const fakeFacilities = () => {
-    // const name : LocaleName = {
-    //     lastName: 'Test',
-    //     firstName: 'Panda',
-    //     middleName: 'The',
-    //     locale: Locale.ENGLISH
-    // }
-    // const degree : Degree = {
-    //     id: '3',
-    //     nameJa: 'パンダの薬',
-    //     nameEn: 'Panda Medicine',
-    //     abbreviation: 'PM'
-    // }
-
-    // const spokenLanguage : SpokenLanguage = {
-    //     iso639_3: 'en-US',
-    //     nameJa: '英語',
-    //     nameEn: 'English',
-    //     nameNative: 'English'
-    // }
-
-    // const specialtyName : SpecialtyName = {
-    //     name: 'Panda Science',
-    //     locale: Locale.ENGLISH
-    // }
-
-    // const specialty : Specialty = {
-    //     id: '3',
-    //     names: [specialtyName]
-    // }
-
-    // const healthcareProfessional : HealthcareProfessional = {
-    //     id: '3',
-    //     names: [name],
-    //     degrees: [degree],
-    //     spokenLanguages: [spokenLanguage],
-    //     specialties: [specialty],
-    //     acceptedInsurance: [Insurance.INTERNATIONAL_HEALTH_INSURANCE]
-    // }
-
     const facility : Facility = {
         id: '1',
         nameEn: 'Zoo',
