+++ src/index.ts
@@ -2,8 +2,9 @@ import { ApolloServer } from '@apollo/server'
 import { startStandaloneServer } from '@apollo/server/standalone'
 import loadSchema from './schema'
 import resolvers from './resolvers'
-import { initializeDb, getFacilities, getHealthcareProfessionals } from './database'
-
+import { initializeDb } from './database'
+import { getFacilityById, getFacilities } from './services/facilityService'
+import { getHealthcareProfessionalById, getHealthcareProfessionals } from './services/healthcareProfessionalService'
 import {seedDatabase} from './databaseSeedTool'
 
 const server = new ApolloServer({
@@ -15,8 +16,9 @@ const server = new ApolloServer({
 async function startServer(port = 3001) {
     await initializeDb()
 
-    // seedDatabase()
-    console.log(await getFacilities())
+    // console.log(await getFacilityById('1'))
+    // console.log(await getHealthcareProfessionalById('1'))
+    // console.log(await getFacilities())
     // console.log(await getHealthcareProfessionals())
 
     await startStandaloneServer(server, {
