+++ src/services/facilityService.ts
@@ -1,5 +1,7 @@
 import { HealthcareProfessional, LocaleName, Degree, 
     Specialty, SpecialtyName, SpokenLanguage, Insurance, Facility } from '../typeDefs/dbSchema'
+import { getFirestore } from 'firebase-admin/firestore'
+import { getHealthcareProfessionalsByIds } from './healthcareProfessionalService'
 
 const tempFirebaseDbGet = () => {
     const name : LocaleName = {
@@ -58,13 +60,56 @@ const tempFirebaseDbGet = () => {
 }
 
 export const getFacilityById = async (id: string) => {
-    const healthPro = tempFirebaseDbGet().find(entity => entity.id == id)
+    const db = getFirestore()
+    const facilityRef = db.collection('facilities')
+    const snapshot = await facilityRef.where('id', '=', id).get()
+    var facilities = []
+    snapshot.forEach(doc => {
+      facilities.push(doc.data())
+    })
 
-    return healthPro
+    for await (const facility of facilities) {
+      await hydrateFacility(facility)
+    }
+
+  return facilities
+}
+
+export async function addFacility(facilityRef, facility) {
+  facilityRef.add(transformFacilityForFirestore(facility))
+}
+
+function transformFacilityForFirestore(facility) {
+  const healthcareProfessionalIds = facility.healthcareProfessionals.map(hp => hp.id)
+  facility.healthcareProfessionals = healthcareProfessionalIds
+
+  return facility
 }
 
 export const getFacilities = async () => {
-    const healthPros = tempFirebaseDbGet()
+    const db = getFirestore()
+    const facilitiesRef = db.collection('facilities')
 
-    return healthPros
+    const snapshot = await facilitiesRef.get()
+    var facilities = []
+    snapshot.forEach(doc => {
+      facilities.push(doc.data())
+    })
+
+    for await (const facility of facilities) {
+      await hydrateFacility(facility)
+    }
+
+    return facilities
+}
+
+async function hydrateFacility(facility) {
+    const db = getFirestore()
+    const hpRef = db.collection('healthcareProfessionals')
+    // const snapshot = await hpRef.where('id', 'in', facility.healthcareProfessionals).get()
+    const healthcareProfessionals = await getHealthcareProfessionalsByIds(facility.healthcareProfessionals)
+
+  facility.healthcareProfessionals = healthcareProfessionals
+
+  return facility
 }
